\name{EPIA}
\alias{EPIA}
\docType{data}
\title{
The Eysenck Personality Inventory Impulsivity Subscale
}
\description{
The data came from a published study and was kindly provided by Dr. Ferrando. A group of 1,033 undergraduate student were asked to check on a 112 mm line segment with two end points (almost never, almost always) using their own judgement for the five items taken from the Spanish version of the EPI-A impulsivity subscale. The direct item score was the distance in mm of the check mark from the left end point (Ferrando, 2002).
}
\usage{data(EPIA)}
\format{
  A data frame with 1033 observations on the following 5 variables.
  \describe{
    \item{\code{V1}}{Item 1}
    \item{\code{V2}}{Item 2}
    \item{\code{V3}}{Item 3}
    \item{\code{V4}}{Item 4}
    \item{\code{V5}}{Item 5}
  }
}
\source{
Ferrando, P.J.(2002). Theoretical and Empirical Comparison between Two Models for Continuous Item Responses. \emph{Multivariate Behavioral Research}, 37(4), 521-542.
}
\references{
Ferrando, P.J.(2002). Theoretical and Empirical Comparison between Two Models for Continuous Item Responses. \emph{Multivariate Behavioral Research}, 37(4), 521-542.
\cr
\cr
\cr
\cr
\cr
}

