\name{print.CRM}
\alias{print.CRM}
\title{
Summary output for the object of class \code{CRM} 
}
\description{
Print the output for the object of class \code{CRM}
}
\usage{
\method{print}{CRM}(x, ...)
}
\arguments{
  \item{x}{
an object of class \code{CRM} obtained from \code{EstCRMitem}
}
\item{...}{
More options to pass to summary and print
}

}

\value{
The summary output includes the information from the last EM Cycle, final parameter estimates, and the descriptive statistics of the original data and the transformed data.
}
