\name{SelfEff}
\alias{SelfEff}
\docType{data}
\title{
Teaching Self Efficacy Scale
}
\description{
The data came from a published study and was kindly provided by Dr. Adnan Kan. A group of 307 pre-service teachers graduated from various departments in the college of education were asked to check on a 11 cm line segment with two end points (can not do at all, highly certain can do) using their own judgement for the 10 items that measure teacher self-efficacy on different activities. The direct item score was the distance in cm of the check mark from the left end point (Kan, 2009).
}
\usage{data(SelfEff)}
\format{
  A data frame with 307 observations on the following 10 items.
  \describe{
    \item{\code{Item1}}{Implementing variety of teaching strategies}
    \item{\code{Item2}}{Creating learning environments for the effective participation of students}
    \item{\code{Item3}}{Providing the acquisition of learning experiences according to the students' physical, emotional, and psychomotor characteristics}
    \item{\code{Item4}}{Creating learning environments in which students can effectively express themselves}
    \item{\code{Item5}}{Implementing a variety of assessment strategies (formal, informal)}
    \item{\code{Item6}}{Understanding differences in students' learning styles and providing learning opportunities in compliance with these differences}
    \item{\code{Item7}}{Using teaching methods and techniques within the framework of educational plans and aims}
    \item{\code{Item8}}{Discovering students' interests and skills and guiding them accordingly }
    \item{\code{Item9}}{Using educational technology effectively in classrooms to support teaching}
    \item{\code{Item10}}{Designing and implementing educational variables such as motivation, reinforcement, feedback, etc.}
  }
}
\source{
Kan, A.(2009). Effect of scale response format on psychometric properties in teaching self-efficacy. \emph{Euroasian Journal of Educational Research}, \emph{34}, 215-228.
}
\references{
Kan, A.(2009). Effect of scale response format on psychometric properties in teaching self-efficacy. \emph{Euroasian Journal of Educational Research}, \emph{34}, 215-228.\cr
}

